re-organized tests
