+++ src/core/__tests__/fixtures/corpus/english.txt
