+++ src/core/__tests__/fixtures/corpus/mat/ugnt.txt
