+++ src/core/__tests__/fixtures/corpus/tit/alignmentData/1.json
