+++ src/core/__tests__/fixtures/corpus/tit/alignmentData/3.json
